Change StringFromInteger to accept long values.

+++ gtk/GUIGTK.cxx
@@ -36,9 +36,9 @@ std::string UTF8FromString(const gui_string &s) {
 	return s;
 }
 
-gui_string StringFromInteger(int i) {
+gui_string StringFromInteger(long i) {
 	char number[32];
-	sprintf(number, "%0d", i);
+	sprintf(number, "%0ld", i);
 	return gui_string(number);
 }
